Combine the different versions of publishing-api
This commit moves all of the methods for  `GdsApi::PublishingApiV2` into
`GdsApi::PublishingApi`

In the process of adding a test helper for path reservation, it became
apparent that the adapters and test helpers for the publishing-api were
in a bit of a mess.

Publishing API has two endpoints, the original and a "V2" one. There has
been some effort to reflect is in gds-api-adapters, but this isn't
always consistent. The path reservation helper was moved to the "V2"
class, but the endpoint it uses is for the original publishing-api
endpoint.

Combining the classes means that the caller doesn't need to worry about
calling the correct version of publishing-api, though the "V2" classes
haven't been removed to make it backwards compatible.
